// angular-groups-list/src/lib/models/group-details.ts
export interface GroupDetails {
    groupId: string;
    groupType: string;
    groupName: string;
    groupLevel: string;
    groupDays: string;
    groupHours: string;
    groupDescription: string;
    groupFirstMeet: string;
    groupLastMeet: string;
    groupLector: string;
    groupLectorFotoName: string;
    groupLectorFotoType: string;
    groupLectorFotoSize: number;
    groupLectorFotoContent: string;
    groupShortName: string;
    groupFreePlaces: number;
}